+++ backend/src/main/java/com/team5/secondhand/chat/chatroom/handler/StompMessageProcessor.java
@@ -3,7 +3,7 @@ package com.team5.secondhand.chat.chatroom.handler;
 import com.team5.secondhand.chat.chatroom.service.ChatroomCacheService;
 import com.team5.secondhand.chat.exception.ErrorType;
 import com.team5.secondhand.chat.notification.service.SessionService;
-import com.team5.secondhand.global.jwt.service.JwtService;
+import com.team5.secondhand.global.auth.JwtService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.Message;
