+++ backend/src/main/java/com/team5/secondhand/global/auth/JwtService.java
@@ -1,10 +1,10 @@
-package com.team5.secondhand.global.jwt.service;
+package com.team5.secondhand.global.auth;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.team5.secondhand.application.member.dto.response.MemberDetails;
 import com.team5.secondhand.global.properties.JwtProperties;
-import com.team5.secondhand.global.jwt.util.JwtUtils;
+import com.team5.secondhand.global.util.JwtUtils;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
